requirements.txt
requests
beautifulsoup4
openai
tiktoken
langchain
tqdm
pinecone-client
python-dotenv